requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
certifi==2022.6.15
charset-normalizer==2.1.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
lxml==4.9.1
multidict==6.0.2
PyMySQL==1.0.2
PyYAML==6.0
requests==2.28.1
SQLAlchemy==1.4.39
tenacity==8.0.1
urllib3==1.26.9
yarl==1.7.2
tqdm~=4.64.0